refactor: resolve #898 - split WebWorkerInterpreter.ts (498 lines) approaching 500-line limit
Extracts workerMessageSenders.ts (106 lines, message response construction) and
workerSharedBufferHandlers.ts (115 lines, shared buffer configuration) from
WebWorkerInterpreter.ts (498→340 lines). Pure refactor, no behavior changes.

Closes #898
